+++ changelog.md
@@ -1,6 +1,7 @@
 # 1.7.5-beta.3
 
 - Fixed not being able to input decimal places in the transition speed customizer UI
+- Fixed crash when pressing the info button on the label format page
 
 # 1.7.5-beta.2